+++ packages/headless-driver-runner-v3/src/platform/graphics/null/NullRenderer.ts
@@ -2,15 +2,15 @@ import { akashicEngine as g } from "@akashic/engine-files";
 
 export class NullRenderer implements g.Renderer {
 	begin(): void {
-		//
+		// do nothing
 	}
 
 	end(): void {
-		//
+		// do nothing
 	}
 
 	clear(): void {
-		//
+		// do nothing
 	}
 
 	drawImage(
@@ -22,7 +22,7 @@ export class NullRenderer implements g.Renderer {
 		_destOffsetX: number,
 		_destOffsetY: number
 	): void {
-		//
+		// do nothing
 	}
 
 	drawSprites(
@@ -35,43 +35,43 @@ export class NullRenderer implements g.Renderer {
 		_canvasOffsetY: number[],
 		_count: number
 	): void {
-		//
+		// do nothing
 	}
 
 	translate(_x: number, _y: number): void {
-		//
+		// do nothing
 	}
 
 	transform(_matrix: number[]): void {
-		//
+		// do nothing
 	}
 
 	opacity(_opacity: number): void {
-		//
+		// do nothing
 	}
 
 	save(): void {
-		//
+		// do nothing
 	}
 
 	restore(): void {
-		//
+		// do nothing
 	}
 
 	fillRect(_x: number, _y: number, _width: number, _height: number, _cssColor: string): void {
-		//
+		// do nothing
 	}
 
 	setCompositeOperation(_operation: g.CompositeOperationString): void {
-		//
+		// do nothing
 	}
 
 	setTransform(_matrix: number[]): void {
-		//
+		// do nothing
 	}
 
 	setOpacity(_opacity: number): void {
-		//
+		// do nothing
 	}
 
 	isSupportedShaderProgram(): boolean {
@@ -79,7 +79,7 @@ export class NullRenderer implements g.Renderer {
 	}
 
 	setShaderProgram(shaderProgram: g.ShaderProgram | null): void {
-		//
+		// do nothing
 	}
 
 	_getImageData(sx: number, sy: number, sw: number, sh: number): g.ImageData {
@@ -95,6 +95,6 @@ export class NullRenderer implements g.Renderer {
 		dirtyWidth?: number,
 		dirtyHeight?: number
 	): void {
-		//
+		// do nothing
 	}
 }